+++ app/services/contact/controller_manager/base_handler.rb
@@ -1,4 +1,4 @@
-class Contact::ControllerManager::Base
+class Contact::ControllerManager::BaseHandler
   attr_reader :owner_mkey, :raw_params, :validation
 
   class << self