+++ spec/services/notification/send_spec.rb
@@ -9,34 +9,32 @@ RSpec.describe Notification::Send do
     # email notification
     #
 
+    before { instance.do }
+
     context 'when email notification' do
       let(:kind) { 'email' }
       subject { notification.reload.state }
 
       context 'when contact owner is existing zazo user with persisted emails' do
         let(:contact) { FactoryGirl.create(:contact, owner_mkey: 'GBAHb0482YxlJ0kYwbIS') }
-        before { instance.do }
 
         it { is_expected.to eq 'sent' }
       end
 
       context 'when contact owner is existing zazo user without persisted emails' do
         let(:contact) { FactoryGirl.create(:contact, owner_mkey: 'XqUn9Fs5YHd75l1rin76') }
-        before { instance.do }
 
         it { is_expected.to eq 'canceled' }
       end
 
       context 'when contact owner is not exist' do
         let(:contact) { FactoryGirl.create(:contact, owner_mkey: 'xxxxxxxxxxxx') }
-        before { instance.do }
 
         it { is_expected.to eq 'canceled' }
       end
 
       context 'when email is not correct' do
         let(:contact) { FactoryGirl.create(:contact, owner_mkey: 'dz4X0EvprPJO6fGysT8X') }
-        before { instance.do }
 
         it { is_expected.to eq 'error' }
       end
@@ -52,7 +50,6 @@ RSpec.describe Notification::Send do
 
       context 'when contact owner is existing zazo push_user' do
         let(:contact) { FactoryGirl.create(:contact, owner_mkey: 'GBAHb0482YxlJ0kYwbIS') }
-        before { instance.do }
 
         it { is_expected.to eq 'sent' }
       end